# Conflicts:
#	app/src/main/java/com/zhiyicx/thinksnsplus/data/beans/UserInfoBean.java
#	app/src/main/java/com/zhiyicx/thinksnsplus/modules/home/message/messagelike/MessageLikeAdapter.java

+++ app/src/main/java/com/zhiyicx/thinksnsplus/modules/dynamic/topdynamic/DynamicTopContract.java
@@ -29,7 +29,7 @@ public interface DynamicTopContract {
     }
 
     interface Presenter extends IBasePresenter {
-        float getBalance();
+        double getBalance();
 
         void stickTop(long feed_id);
     }
